♻️ Refactors events models to shared dir
Moves events and organisation models to the `_pydantic/models` directory.
