+++ functions/_pydantic/shared_models.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  <stdin>
-#   timestamp: 2025-07-22T21:15:51+00:00
+#   timestamp: 2025-07-22T21:26:54+00:00
 
 from __future__ import annotations
