+++ functions/openapi.json
@@ -25,7 +25,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventObject": {
         "type": "object",
@@ -109,7 +109,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventObjectPublic": {
         "type": "object",
@@ -185,7 +185,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "CreateEventRequest": {
         "type": "object",
@@ -198,7 +198,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "UpdateEventRequest": {
         "type": "object",
@@ -211,7 +211,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventResponse": {
         "type": "object",
@@ -221,7 +221,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventListResponse": {
         "type": "object",
@@ -234,7 +234,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventResponsePublic": {
         "type": "object",
@@ -244,7 +244,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "EventListResponsePublic": {
         "type": "object",
@@ -257,7 +257,7 @@
           }
         },
         "x-internal-group": "events",
-        "x-internal-path": "functions/events"
+        "x-internal-path": "_pydantic/models"
       },
       "OrganisationObject": {
         "type": "object",
@@ -524,7 +524,7 @@
       "get": {
         "summary": "Trigger an event",
         "description": "Trigger and event.",
-        "operationId": "eventbridge-1566a242-991f-45b1-b822-9e1a26b0f492",
+        "operationId": "eventbridge-04b0e752-74ca-42ae-a4ca-17e769d6d1ad",
         "parameters": [],
         "tags": [
           "Event_Brdige"
@@ -596,7 +596,7 @@
       "post": {
         "summary": "Create Event",
         "description": "Creates a new event and stores its location.",
-        "operationId": "events-ef394575-effe-44e9-b874-3c16cd6a1224",
+        "operationId": "events-399b9794-ff44-4c19-8714-fa942f93c77f",
         "parameters": [
           {
             "name": "organisation",
@@ -653,7 +653,7 @@
       "get": {
         "summary": "Get Single Event",
         "description": "Fetch details of a specific event by its KSUID.",
-        "operationId": "events-baf3b5b6-6a8d-4838-a656-c60bd4b01bf5",
+        "operationId": "events-36f07e95-df90-424c-a929-86ea0fc99e2a",
         "parameters": [
           {
             "name": "organisation",
@@ -709,7 +709,7 @@
       "put": {
         "summary": "Update an Event",
         "description": "Update an existing event and its location",
-        "operationId": "events-72436075-5a5f-4833-994d-a3d449514117",
+        "operationId": "events-cffedcfe-f909-453a-83c2-959a2c0f31e1",
         "parameters": [
           {
             "name": "organisation",
@@ -775,7 +775,7 @@
       "get": {
         "summary": "Public Get Event",
         "description": "Public endpoint to get a specific event by its KSUID.",
-        "operationId": "events-5eb4ba42-30e0-433b-a55e-77c42ec56aa8",
+        "operationId": "events-df05da30-ede1-40f9-822e-0f82d3068bad",
         "parameters": [
           {
             "name": "organisation",
@@ -837,7 +837,7 @@
       "get": {
         "summary": "Public Get ALL Events",
         "description": "Public endpoint to get a specific event by its KSUID.",
-        "operationId": "events-b4af1ae6-2bcd-45af-9be7-ae19ae9c4041",
+        "operationId": "events-c16f81f9-dee2-46f0-bfc0-df272e290d49",
         "parameters": [
           {
             "name": "organisation",
@@ -936,7 +936,7 @@
       "post": {
         "summary": "Generate Presigned Download URL",
         "description": "Generates a presigned GET URL to allow users to download a file from S3.",
-        "operationId": "presigned_url-5525b61b-011f-4c64-ac0d-f122275cb81a",
+        "operationId": "presigned_url-61c25f85-c9fd-41e9-a073-5210c73bcc06",
         "parameters": [],
         "tags": [
           "S3"
@@ -1005,7 +1005,7 @@
       "post": {
         "summary": "Create Customer",
         "description": "Creates a new customer record under a specific organisation.",
-        "operationId": "customers-b39109bc-96a1-4b2c-a1f7-21f69c1d8056",
+        "operationId": "customers-df09d37e-1a60-4481-9003-7fe52644172a",
         "parameters": [
           {
             "name": "organisation",
@@ -1051,7 +1051,7 @@
       "get": {
         "summary": "Get Single Customer",
         "description": "Fetch details of a specific customer by their KSUID in a given organisation.",
-        "operationId": "customers-67091ccd-ca78-4aeb-af54-25b6be18410f",
+        "operationId": "customers-042f11cf-e972-4560-a81a-770477f982d9",
         "parameters": [
           {
             "name": "organisation",
@@ -1171,7 +1171,7 @@
       "put": {
         "summary": "Update Organisatioon Settings",
         "description": "Fetch details of a specific organisation by its slug.",
-        "operationId": "organisation-49b644b1-a69c-4091-b9ae-ecbcbad7b3db",
+        "operationId": "organisation-35041ae1-8d17-4eda-9505-723412cc3152",
         "parameters": [
           {
             "name": "organisation",
